refactored exercises
