--- app/(member-area)/profile/exercises/create/navbar.tsx
@@ -1,12 +1,0 @@
-import BackButton from "@/components/back-button";
-
-export default function Navbar() {
-  return (
-    <>
-      <div className="grid grid-cols-3 items-center py-3">
-        <BackButton />
-        <span className="text-center">Create Exercise</span>
-      </div>
-    </>
-  );
-}
